docs: add requirements and update makefile cmd list example
