+++ cmd/admin/static/templates/dashboard.page.gohtml
