+++ cmd/admin/static/templates/new-password.page.gohtml
