--- cmd/adminapi/README.md
@@ -1,1 +1,0 @@
-# Admin API
