+++ internal/admin/webpage/auth.go
@@ -13,8 +13,6 @@ func (page *WebPage) Login(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-func (page *WebPage) Logout(w http.ResponseWriter, r *http.Request) {}
-
 func (page *WebPage) ForceNewPassword(w http.ResponseWriter, r *http.Request) {
 	if err := page.render.Template(w, r, "new-password", nil); err != nil {
 		page.logger.Error("new-password", zap.Error(err))
