--- internal/adminclient/middleware.go
@@ -1,20 +1,0 @@
-package adminclient
-
-import (
-	"net/http"
-)
-
-// SessionLoad middleware automatically loads and saves session data for the current request,
-// and communicates the session token to and from the client in a cookie.
-func SessionLoad(next http.Handler) http.Handler {
-	return session.LoadAndSave(next)
-}
-
-func withAuth(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if !session.Exists(r.Context(), "userID") {
-			http.Redirect(w, r, "/login", http.StatusTemporaryRedirect)
-		}
-		next.ServeHTTP(w, r)
-	})
-}
